+++ TWiT.tv/TWShowViewController.m
@@ -112,9 +112,9 @@
         self.emailButton.hidden = !self.show.email;
         self.phoneButton.hidden = !self.show.phone;
 
-        for (int i = 0; i < [self.tableView numberOfSections]; i++)
+        for (NSInteger i = 0; i < [self.tableView numberOfSections]; i++)
         {
-            for (int j = 0; j < [self.tableView numberOfRowsInSection:i]; j++)
+            for (NSInteger j = 0; j < [self.tableView numberOfRowsInSection:i]; j++)
             {
                 NSIndexPath *indexPath = [NSIndexPath indexPathForRow:j inSection:i];
                 TWEpisodeCell *cell = (TWEpisodeCell*)[self.tableView cellForRowAtIndexPath:indexPath];
@@ -143,7 +143,7 @@
 
 - (IBAction)openDetailView:(UIButton*)sender
 {
-    float headerHeight = self.tableView.tableHeaderView.frame.size.height;
+    CGFloat headerHeight = self.tableView.tableHeaderView.frame.size.height;
     
     if(self.tableView.contentOffset.y <= -self.view.bounds.size.height+headerHeight)
     {
@@ -235,7 +235,7 @@
 - (void)scrollViewDidScroll:(UIScrollView*)scrollView
 {
     UIAccessibilityPostNotification(UIAccessibilityLayoutChangedNotification, nil);
-    float headerHeight = self.tableView.tableHeaderView.frame.size.height;
+    CGFloat headerHeight = self.tableView.tableHeaderView.frame.size.height;
     
     if(scrollView == self.tableView)
     {
@@ -306,7 +306,7 @@
     TWEpisodeCell *cell = (TWEpisodeCell*)[self.tableView cellForRowAtIndexPath:indexPath];
     
     if([notification.name isEqualToString:@"enclosureDownloadDidReceiveData"])
-        cell.progress = (enclosure.expectedLength != 0)? enclosure.downloadedLength/(float)enclosure.expectedLength : 0;
+        cell.progress = (enclosure.expectedLength != 0)? enclosure.downloadedLength/(CGFloat)enclosure.expectedLength : 0;
     else if([notification.name isEqualToString:@"enclosureDownloadDidFinish"] || [notification.name isEqualToString:@"enclosureDownloadDidFail"])
         cell.progress = 1;
 }
